pyproject.toml
[tool.poetry]
name = "lucidtask"
version = "0.1.0"
description = "A FastAPI , Sqlalchemy Application for submission task"
authors = ["Atabak Hooshangi", "1990.atabak.hooshangi@gmail.com"]
license = "MIT"
readme = "README.md"
keywords = ["fastapi", "web", "api"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
"annotated-types" = "0.6.0"
anyio = "4.3.0"
click = "8.1.7"
colorama = "0.4.6"
fastapi = "0.109.2"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
pydantic = "2.6.1"
"pydantic-core" = "2.16.2"
"python-dotenv" = "1.0.1"
PyYAML = "6.0.1"
sniffio = "1.3.0"
starlette = "0.36.3"
"typing-extensions" = "4.9.0"
uvicorn = "0.27.1"
watchfiles = "0.21.0"
websockets = "12.0"
graphene = "3.3"
asyncpg = "0.29.0"
alembic = "1.13.1"
starlette-graphene3 = "^0.6.0"
email-validator = "^2.1.0.post1"
sqlalchemy = "^2.0.27"
psycopg = "^3.1.18"
pydantic-settings = "^2.2.1"
psycopg-binary = "^3.1.18"
itsdangerous = "^2.1.2"
orjson = "^3.9.15"
faker = "^23.3.0"

pyjwt = "^2.8.0"
aiomysql = "^0.2.0"
asyncmy = "^0.2.9"
pymysql = "^1.1.1"
cryptography = "^42.0.8"
bcrypt = "^4.1.3"
[tool.poetry.dev-dependencies]
pytest = "^6.2.5"